+++ rollup.config.js
@@ -20,6 +20,7 @@ function getRollupObject ({minifying, format = 'umd'} = {}) {
   return nonMinified;
 }
 
+// eslint-disable-next-line import/no-anonymous-default-export
 export default [
   getRollupObject(),
   getRollupObject({minifying: true}),